Created various classes and dynamically composed some components.
The main changes made were the creation of the IssueList, Issuetable, and
Issue Row classes with simple structures (IssueAdd and IssueFilter skeletons
were also created). A global array was also introduced to simulate an
array retrieved from a server. The array was passed from IssueList to
IssueTable where the objects in the array were mapped to an array of
IssueRow instances. The components were nested and then rendered.